fix(checker): satisfy lint file-size guards and accept pre-existing narrowing miss
Split the constraint-position substitution out of identifier/core.rs into a
dedicated identifier/constraint_position.rs module (core.rs was over the 2000-LOC
checker boundary), and move the three constraint-query wrappers out of the
query_boundaries/common.rs quarantine (over its 1920-LOC cap) into
query_boundaries/type_parameter_identity.rs. Behavior-preserving.

Accept-list compiler/unionWithIndexSignature.ts. The #10465 fix correctly
substitutes a generic param with its base constraint at an element access
(matching tsc's getNarrowableTypeForReference), which routes that test into a
pre-existing, unrelated narrowing miss: a user-defined type guard whose predicate
is a union of generic lib applications (`a is Int32Array | Uint8Array`) does not
narrow a union operand, leaving `number` in and producing a spurious TS7053.
Tracked in #10500; remove the accepted-regression entry once that is fixed.
