+++ crates/tsz-checker/src/types/computation/identifier/mod.rs
@@ -4,6 +4,7 @@
 //! the binder, checking TDZ violations, validating definite assignment,
 //! applying flow-based narrowing, and handling intrinsic/global names.
 
+mod constraint_position;
 mod core;
 mod resolution;
